feat(initial): Initial version
This is the first version of the project - it does not yet publish to
MQTT, but it does publish to InfluxDB which was my primary initial use
case. The only metrics published are those that are gathered from the
dual plantower sensors. Metrics like temperature, humidity, etc, are
not yet published. Also, this is embarassingly bad code.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>