feat(mqtt): actually publish to MQTT
For a program names purpleair2mqtt, you'd think this would've actually
published to MQTT, but nope, it wasn't actually doing that. This adds in
the code to allow it to publish to MQTT, but does not yet publish to
allow for HomeAssistant discovery.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>